💄 Style: 파동, 배경화면 범위 화면 전체로 수정
