+++ src/features/multi-step-form/ui/careerForm/secondStep.tsx
@@ -2,11 +2,13 @@
 
 import { useShallow } from 'zustand/shallow';
 
-import Ball from '@/features/multi-step-form/ui/careerForm/tagFormStep/ui/Ball';
+import { cn } from '@/shared/lib/utils';
+import { spacingStyles } from '@/shared/spacing';
 import { useCardFormStore } from '@/shared/store/cardFormState';
 import { Button } from '@/shared/ui/button';
 import Header from '@/shared/ui/header';
 
+import { Ball } from './tagFormStep/ui/Ball';
 import TagBox from './tagFormStep/ui/TagBox';
 
 type SecondStepProps = {
@@ -30,12 +32,39 @@ function SecondStep({ handleNextStep }: SecondStepProps) {
     resetTagCount();
   };
 
+  const circleStyles = ['absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2  overflow-hidden rounded-full'];
+
+  const shadowStyles = [
+    'shadow-[0px_0px_11.9px_rgba(103,68,255,0.2),_inset_0px_0px_54.5px_rgba(208,195,255,0.2)]',
+    'shadow-[0px_0px_11.9px_rgba(103,68,255,0.3),_inset_0px_0px_54.5px_rgba(208,195,255,0.3)]',
+  ];
   return (
     <>
       <div className="h-[72dvh] w-full items-center justify-start gap-4 overflow-hidden">
-        <main className="relative flex h-full w-full flex-col items-center justify-between">
+        <main
+          className={cn(
+            spacingStyles({ paddingTop: 'lg' }),
+            `relative flex h-full w-full flex-col items-center justify-between`,
+          )}
+        >
           <Header title={`명함에 추가할 태그를 \n 최소 2개 이상 선택해 주세요`} />
           <Ball tagCount={tagCount} />
+          <div
+            className={cn(
+              'pointer-events-none z-10 h-[494px] w-[494px] opacity-40',
+              circleStyles.join(' '),
+              shadowStyles[0],
+              'firstWave',
+            )}
+          />
+          <div
+            className={cn(
+              'pointer-events-none z-20 h-72 w-72 opacity-50',
+              circleStyles.join(' '),
+              shadowStyles[1],
+              'secondWave',
+            )}
+          />
           <div className="relative h-full w-full">
             <TagBox />
           </div>
