+++ src/shared/ui/appbar.tsx
@@ -14,7 +14,7 @@ const appbarVariants = cva(
       page: {
         main: '',
         detail: '',
-        create: 'bg-gray-black',
+        create: '',
         mypage: 'bg-core-black',
         received: '',
       },
@@ -45,6 +45,7 @@ type AppbarProps = AppbarVariantProps & {
   title?: string;
   router?: () => void;
   isBlurred?: boolean;
+  className?: string;
 };
 
 function renderLeftIcon({
@@ -166,13 +167,12 @@ function Appbar({
   onRightClickSecond,
   router,
   isBlurred,
+  className,
 }: AppbarProps) {
   return (
-    <header className={cn(appbarVariants({ page, hasBackground }))}>
+    <header className={cn(className, appbarVariants({ page, hasBackground }))}>
       <div className="flex flex-1">{renderLeftIcon({ page, onLeftClick, isBlurred })}</div>
-
       {title && <h1 className="flex-1 text-center text-body-3 text-white">{title}</h1>}
-
       <div className="flex flex-1 justify-end">
         {renderRightIcon({ page, onRightClick, onRightClickSecond, router, isBlurred })}
       </div>
